Renamed specs into versioned format
- latest -> v2
- canary -> v3
- latest & canary are now acting as aliases to versions instead of "defining" them
