+++ lib/specs/index.js
@@ -1,5 +1,13 @@
 module.exports = {
     get: function get(key) {
-        return require(`./${key}`);
+        let [version] = key;
+
+        if (version === 'latest') {
+            version = 'v2';
+        } else if (version === 'canary') {
+            version = 'v3';
+        }
+
+        return require(`./${[version]}`);
     }
 };
