+++ lib/specs/v3.js
@@ -1,8 +1,8 @@
 const _ = require('lodash');
-const previousSpec = require('./latest');
+const previousSpec = require('./v2');
 const ghostVersions = require('../utils').versions;
 const docsBaseUrl = `https://ghost.org/docs/api/handlebars-themes/`;
-const prevDocsBaseUrl = `https://themes.ghost.org/v${ghostVersions.latest.docs}/docs/`;
+const prevDocsBaseUrl = `https://themes.ghost.org/v${ghostVersions.v2.docs}/docs/`;
 const prevDocsBaseUrlRegEx = new RegExp(prevDocsBaseUrl, 'g');
 
 const previousKnownHelpers = previousSpec.knownHelpers;
